TEST-0 finish

+++ Figure.Lib/Figures/IFigure.cs
@@ -2,5 +2,5 @@
 
 public interface IFigure
 {
-    decimal GetArea();
+    double GetArea();
 }
